+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes
 
+## 1.2
+
+* `repeat` argument added to `Imposter.add_stub`
+* Multiple responses can be added to the same stub
+
 ## 1.1
 
 * Imposter.mockhttp by default waits for mountebank to become accessible